+++ sisl/io/siesta/out.py
@@ -672,7 +672,7 @@ class outSileSiesta(SileSiesta):
         return val
 
     @sile_fh_open()
-    def read_scf(self, key="scf", iscf=-1, imd=None, as_dataframe=False):
+    def read_scf(self, key="scf", iscf=-1, imd=None, as_dataframe=False, ret_header=False):
         r""" Parse SCF information and return a table of SCF information depending on what is requested
 
         Parameters
@@ -690,6 +690,9 @@ class outSileSiesta(SileSiesta):
             whether the information should be returned as a `pandas.DataFrame`. The advantage of this
             format is that everything is indexed and therefore you know what each value means.You can also
             perform operations very easily on a dataframe. 
+        ret_header: bool, optional
+            whether to also return the headers that define each value in the returned array,
+            will have no effect if `as_dataframe` is true.
         """
 
         #These are the properties that are written in SIESTA scf
@@ -710,14 +713,31 @@ class outSileSiesta(SileSiesta):
 
         def common_parse(line, d):
             if line.startswith('ts-Vha:'):
-                d['ts-Vha'] = float(line.split()[1])
+                d['ts-Vha'] = [float(line.split()[1])]
+                if 'ts-Vha' not in props:
+                    d['order'].append("ts-Vha")
+                    props.append("ts-Vha")
+            elif line.startswith("spin moment: S"):
+                # 4.1 and earlier
+                d['S'] = list(map(float, line.split("=")[1].split()[1:]))
+                if 'Sx' not in props:
+                    d['order'].append("S")
+                    props.extend(['Sx', 'Sy', 'Sz'])
+            elif line.startswith("spin moment: {S}"):
+                # 4.2 and later
+                d['S'] = list(map(float, line.split("= {")[1].split()[:3]))
+                if 'Sx' not in props:
+                    d['order'].append("S")
+                    props.extend(['Sx', 'Sy', 'Sz'])
             elif line.startswith('bulk-bias: |v'):
                 d['bb-v'] = list(map(float, line.split()[-3:]))
-                if 'bb-vx' not in props:
+                if 'BB-vx' not in props:
+                    d['order'].append("bb-v")
                     props.extend(['BB-vx', 'BB-vy', 'BB-vz'])
             elif line.startswith('bulk-bias: {q'):
                 d['bb-q'] = list(map(float, line.split()[-3:]))
-                if 'bb-q+' not in props:
+                if 'BB-q+' not in props:
+                    d['order'].append("bb-q")
                     props.extend(['BB-q+', 'BB-q-', 'BB-q0'])
             else:
                 return False
@@ -745,10 +765,9 @@ class outSileSiesta(SileSiesta):
                         # Populate DATA by splitting
                         data = line.split()
                         data =  [int(data[1])] + list(map(float, data[2:]))
-                    d['data'] = data
+                    construct_data(d, data)
 
         elif key.lower() == 'ts-scf':
-            props.append("ts-Vha")
             def parse_next(line, d):
                 line = line.strip().replace('*', '0')
                 reset_d(d, line)
@@ -769,23 +788,29 @@ class outSileSiesta(SileSiesta):
                     if len(line) == 100:
                         data = [int(line[8:12]), float(line[12:28]), float(line[28:44]),
                                 float(line[44:60]), float(line[60:70]), float(line[70:80]),
-                                float(line[80:90]), float(line[90:100]), d['ts-Vha']] + d['ts-q']
+                                float(line[80:90]), float(line[90:100])]
                     elif len(line) == 90:
                         data = [int(line[8:12]), float(line[12:28]), float(line[28:44]),
                                 float(line[44:60]), float(line[60:70]), float(line[70:80]),
-                                float(line[80:90]), d['ts-Vha']] + d['ts-q']
+                                float(line[80:90])]
                     else:
                         # Populate DATA by splitting
                         data = line.split()
-                        data = [int(data[1])] + list(map(float, data[2:])) + [d['ts-Vha']] + d['ts-q']
-                    d['data'] = data
+                        data = [int(data[1])] + list(map(float, data[2:]))
+                    construct_data(d, data)
 
         # A temporary dictionary to hold information while reading the output file
         d = {
             '_found_iscf': False,
             '_final_iscf': 0,
             'data': [],
+            'order': [],
         }
+        def construct_data(d, data):
+            for key in d["order"]:
+                data.extend(d[key])
+            d["data"] = data
+
         md = []
         scf = []
         for line in self:
@@ -873,11 +898,13 @@ class outSileSiesta(SileSiesta):
                     df.reset_index("iscf", inplace=True)
                 return df
 
-            if iscf is None:
+            if iscf is not None:
                 # since each MD step may be a different number of SCF steps
-                # we cannot convert to a dense array
-                return md
-            return np.array(md)
+                # we can only convert for a specific entry
+                md = np.array(md)
+            if ret_header:
+                return md, props
+            return md
 
         # correct imd to ensure we check against the final size
         imd = min(len(md) - 1, max(len(md) + imd, 0))
@@ -886,7 +913,10 @@ class outSileSiesta(SileSiesta):
             if as_dataframe:
                 return pd.DataFrame(index=pd.Index([], name="iscf"),
                                     columns=props[1:])
-            return np.array(md[imd])
+            md = np.array(md[imd])
+            if ret_header:
+                return md, props
+            return md
 
         if imd > len(md):
             raise ValueError(f"{self.__class__.__name__}.read_scf could not find requested MD step ({imd}).")
@@ -896,6 +926,8 @@ class outSileSiesta(SileSiesta):
         scf = np.array(md[imd])
         if as_dataframe:
             return MDstep_dataframe(scf)
+        if ret_header:
+            return scf, props
         return scf
 
     @sile_fh_open()